+++ common/constants.h
@@ -17,7 +17,7 @@
 #define SW_VERSION_MAJOR 1
 #define SW_VERSION_MINOR 0
 #define SW_VERSION_PATCH 6
-#define SW_VERSION_BUILD 0
+#define SW_VERSION_BUILD 3
 
 // Number of max sensors connected to the board
 #define CONNECTED_SENSORS 8