+++ sensors/sunlight/sunlight.c
@@ -357,8 +357,6 @@ int sunlight_init(sensor_t* sunlight, sensor_config_t* config)
         sl_power(sunlight, false);
         return ret; 
     }
-
-    sunlight_reset(); // Reset sensor
     
     sl_power(sunlight, false); // Power off
     if (!ret)
@@ -379,39 +377,31 @@ int sunlight_read_config(sensor_config_t* config)
     //sunlight_write(REG_CLEAR_ERROR, &data, 1); // Wake sensor up
 
     if ((ret = sunlight_read(REG_METER_CONTROL, &data, 1)) != 0) return ret; // Read meter control register
-    config->enable_nRDY = data & 0b1 << 0; // Save meter control data
-    config->enable_abc = data & 0b1 << 1;
-    config->enable_static_IIR = data & 0b1 << 2;
-    config->enable_dynamic_IIR = data & 0b1 << 3;
-    config->enable_pressure_comp = data & 0b1 << 4;
-    config->invert_nRDY = data & 0b1 << 5;
+    config->enable_nRDY = !(data & 0b1 << 0); // Save meter control data
+    config->enable_abc = !(data & 0b1 << 1);
+    config->enable_static_IIR = !(data & 0b1 << 2);
+    config->enable_dynamic_IIR = !(data & 0b1 << 3);
+    config->enable_pressure_comp = !(data & 0b1 << 4);
+    config->invert_nRDY = !(data & 0b1 << 5);
 
     if ((ret = sunlight_read(REG_MEAS_MODE, buf, 13)) != 0) return ret; // Read measurement registers
     config->single_meas_mode = (bool)buf[0]; // Save measurement settings data
-    // config->meas_period = ntoh16(*( (uint16_t*)&buf[1]));
     memcpy(&config->meas_period, &buf[1], 2);
     config->meas_period = ntoh16(config->meas_period);
-    // config->meas_samples = ntoh16(*( (uint16_t*)&buf[3]));
     memcpy(&config->meas_samples, &buf[3], 2);
     config->meas_samples = ntoh16(config->meas_samples);
-    // config->abc_period = ntoh16(*( (uint16_t*)&buf[5]));
     memcpy(&config->abc_period, &buf[5], 2);
     config->abc_period = ntoh16(config->abc_period);
-    // config->abc_target_value = ntoh16(*( (uint16_t*)&buf[9]));
     memcpy(&config->abc_target_value, &buf[9], 2);
     config->abc_target_value = ntoh16(config->abc_target_value);
     config->filter_coeff = (uint16_t)buf[12];
-
-    if ((ret = sunlight_read(REG_AIR_PRESSURE_H, buf, 2)) != 0) return ret; // Read pressure register
-    // config->pressure = ntoh16(*( (uint16_t*)&buf[0])) / 10;
-    memcpy(&config->pressure, &buf[0], 2);
-    config->pressure = ntoh16(config->pressure) / 10;
 
     return SUCCESS;
 }
 
 static int sl_write_config(sensor_config_t* config)
 {
+    bool changed = false;
     int32_t ret;
     sensor_config_t read_config;
     if ((ret = sunlight_read_config(&read_config)) != 0) return ret; // Read SUNRISE config
@@ -424,13 +414,14 @@ static int sl_write_config(sensor_config_t* config)
         read_config.invert_nRDY != config->invert_nRDY)
     {
         print_ser_output(SEVERITY_WARN, SOURCE_SENSORS, SOURCE_SUNLIGHT, "Config - Writing meter control");
+        changed = true;
         uint8_t data = 0; // meter control vector
-        data |= config->enable_nRDY;
-        data |= config->enable_abc << 1;
-        data |= config->enable_static_IIR << 2;
-        data |= config->enable_dynamic_IIR << 3;
-        data |= config->enable_pressure_comp << 4;
-        data |= config->invert_nRDY << 5;
+        data |= !config->enable_nRDY;
+        data |= !config->enable_abc << 1;
+        data |= !config->enable_static_IIR << 2;
+        data |= !config->enable_dynamic_IIR << 3;
+        data |= !config->enable_pressure_comp << 4;
+        data |= !config->invert_nRDY << 5;
 
         if ((ret = sunlight_write(REG_METER_CONTROL, &data, 1)) != 0) return ret; // Write measurement control register
     }
@@ -442,18 +433,15 @@ static int sl_write_config(sensor_config_t* config)
         read_config.filter_coeff != config->filter_coeff)
     {
         print_ser_output(SEVERITY_WARN, SOURCE_SENSORS, SOURCE_SUNLIGHT, "Config - Writing full configuration");
+        changed = true;
         uint8_t command_buf[13] = {0};
         command_buf[0] = (uint8_t)config->single_meas_mode; // Prepare command buffer
-        // *( (uint16_t*)&command_buf[1]) = ntoh16(config->meas_period);
         uint16_t val = ntoh16(config->meas_period);
         memcpy(&command_buf[1], &val, 2);
-        // *( (uint16_t*)&command_buf[3]) = ntoh16(config->meas_samples);
         val = ntoh16(config->meas_samples);
         memcpy(&command_buf[3], &val, 2);
-        // *( (uint16_t*)&command_buf[5]) = ntoh16(config->abc_period);
         val = ntoh16(config->abc_period);
         memcpy(&command_buf[5], &val, 2);
-        // *( (uint16_t*)&command_buf[9]) = ntoh16(config->abc_target_value);
         val = ntoh16(config->abc_target_value);
         memcpy(&command_buf[9], &val, 2);
 
@@ -461,16 +449,22 @@ static int sl_write_config(sensor_config_t* config)
 
         if ((ret = sunlight_write(REG_MEAS_MODE, command_buf, 13)) != 0) return ret; // Write measurement registers
     }
-
-    if (read_config.pressure != config->pressure && config->enable_pressure_comp) // Check pressure
+    if (changed) // If configuration changed
     {
-        print_ser_output(SEVERITY_WARN, SOURCE_SENSORS, SOURCE_SUNLIGHT, "Config - Writing pressure");
-        uint8_t command_buf[2] = {0};
-        // *( (uint16_t*)&command_buf[0]) = ntoh16(config->pressure * 10); // Prepare command
-        uint16_t val = ntoh16(config->pressure * 10); // Prepare command
-        memcpy(&command_buf[0], &val, 2);
-
-        if ((ret = sunlight_write(REG_AIR_PRESSURE_H, command_buf, 2)) != 0) return ret; // Write pressure register
+        sunlight_reset();
+    }
+    if (config->enable_pressure_comp) // If pressure compensation enabled write pressure from EEPROM
+    {
+        uint8_t buf[2];
+        uint16_t val = ntoh16(config->pressure * 10);
+        memcpy(buf, &val, 2);
+        if ((ret = sunlight_write(REG_AIR_PRESSURE_H, buf, 2)) != 0) return ret;
+        sleep_ms(10);
+        uint8_t recv[2];
+        if ((ret = sunlight_read(REG_AIR_PRESSURE_H, recv, 2)) != 0) return ret;
+        memcpy(&val, recv, 2);
+        if (config->pressure != ntoh16(val) / 10) 
+            print_ser_output(SEVERITY_ERROR, SOURCE_SENSORS, SOURCE_SUNLIGHT, "Pressure mismatch: Written: %i, read: %i", config->pressure, ntoh16(val) / 10);
     }
     return SUCCESS;
 }